refactor: relocate remote-shell env helpers into engine-primitives
remote_shell_path_preamble and quote_runner_env_value are pure shell-command
construction, not runner behavior. They lived in runner/command_path.rs, which
forced server/client/ssh_client.rs (a core module) to depend on crate::runner.

Move both (plus the private escape_double_quoted_env_value helper) into
homeboy-engine-primitives::shell alongside quote_arg/quote_args, and repoint
ssh_client at crate::engine::shell. This severs one core->runner edge with no
hook and no type-drag, advancing the runner-crate extraction.
